RNA-seq: refactor the 'bowtie_mapping_stats.py' utility to put into bcftbx.cli.
