+++ bin/bowtie_mapping_stats.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+#
+#     bowtie_mapping_stats.py: write Bowtie mapping stats to a spreadsheet
+#     Copyright (C) University of Manchester 2021 Peter Briggs
+#
+from bcftbx.cli.bowtie_mapping_stats import main
+if __name__ == "__main__":
+     main()
